Common: merged path_ex.cpp back into path.cpp
As a consequence this requires Tools to link few Allegro sources:
- allegro.c - contains std lib implementations (alloc etc)
- file.c
- wfile.c (on Windows)
- ufile.c (elsewhere
- unicode.c

All above are added to TOOLS_MINIMAL_ALLEGRO_SOURCES in CMakeLists.

MSVS specific: Allegro's file.c must be configured to output as al_file.obj because Common lib already has a file.cpp, which output obj will conflict with allegro's.
