+++ Common/util/path.cpp
@@ -11,12 +11,13 @@
 // https://opensource.org/license/artistic-2-0/
 //
 //=============================================================================
+#include "util/path.h"
+#include <allegro/file.h>
 #include "platform/platform.h"
 #if AGS_PLATFORM_OS_WINDOWS
 #include "platform/windows/windows.h"
 #endif
 #include "util/file.h"
-#include "util/path.h"
 #include "util/stdio_compat.h"
 
 namespace AGS
@@ -81,6 +82,24 @@ String ReplaceExtension(const String &filename, const String &ext)
     return String::FromFormat("%s.%s", noext.GetCStr(), ext.GetCStr());
 }
 
+int ComparePaths(const String &path1, const String &path2)
+{
+    // Make minimal absolute paths
+    String fixed_path1 = MakeAbsolutePath(path1);
+    String fixed_path2 = MakeAbsolutePath(path2);
+
+    fixed_path1.TrimRight('/');
+    fixed_path2.TrimRight('/');
+
+    int cmp_result =
+#if defined AGS_CASE_SENSITIVE_FILESYSTEM
+        fixed_path1.Compare(fixed_path2);
+#else
+        fixed_path1.CompareUtf8NoCase(fixed_path2);
+#endif // AGS_CASE_SENSITIVE_FILESYSTEM
+    return cmp_result;
+}
+
 String GetDirectoryPath(const String &path)
 {
     if (File::IsDirectory(path))
@@ -95,6 +114,28 @@ String GetDirectoryPath(const String &path)
         return dir;
     }
     return "./";
+}
+
+bool IsSameOrSubDir(const String &parent, const String &path)
+{
+    char can_parent[MAX_PATH_SZ];
+    char can_path[MAX_PATH_SZ];
+    char relative[MAX_PATH_SZ];
+    // canonicalize_filename treats "." as "./." (file in working dir)
+    const char *use_parent = parent == "." ? "./" : parent.GetCStr();
+    const char *use_path   = path   == "." ? "./" : path.GetCStr();
+    canonicalize_filename(can_parent, use_parent, MAX_PATH_SZ);
+    canonicalize_filename(can_path, use_path, MAX_PATH_SZ);
+    const char *pstr = make_relative_filename(relative, can_parent, can_path, MAX_PATH_SZ);
+    if (!pstr)
+        return false;
+    for (pstr = strstr(pstr, ".."); pstr && *pstr; pstr = strstr(pstr, ".."))
+    {
+        pstr += 2;
+        if (*pstr == '/' || *pstr == '\\' || *pstr == 0)
+            return false;
+    }
+    return true;
 }
 
 bool IsRelativePath(const String &path)
@@ -160,6 +201,36 @@ String MakeTrailingSlash(const String &path)
     String dir_path = String::FromFormat("%s/", path.GetCStr());
     FixupPath(dir_path);
     return dir_path;
+}
+
+String MakeAbsolutePath(const String &path)
+{
+    if (path.IsEmpty())
+    {
+        return "";
+    }
+    // canonicalize_filename treats "." as "./." (file in working dir)
+    String abs_path = path == "." ? "./" : path;
+    char buf[MAX_PATH_SZ];
+    canonicalize_filename(buf, abs_path.GetCStr(), MAX_PATH_SZ);
+    abs_path = buf;
+    FixupPath(abs_path);
+    return abs_path;
+}
+
+String MakeRelativePath(const String &base, const String &path)
+{
+    char can_parent[MAX_PATH_SZ];
+    char can_path[MAX_PATH_SZ];
+    char relative[MAX_PATH_SZ];
+    // canonicalize_filename treats "." as "./." (file in working dir)
+    const char *use_parent = base == "." ? "./" : base.GetCStr();
+    const char *use_path = path == "." ? "./" : path.GetCStr(); // FIXME?
+    canonicalize_filename(can_parent, use_parent, MAX_PATH_SZ);
+    canonicalize_filename(can_path, use_path, MAX_PATH_SZ);
+    String rel_path = make_relative_filename(relative, can_parent, can_path, MAX_PATH_SZ);
+    FixupPath(rel_path);
+    return rel_path;
 }
 
 String &AppendPath(String &path, const String &child)
