+++ Solutions/Common.Lib/Common.Lib.vcxproj
@@ -696,7 +696,6 @@
     <ClCompile Include="..\..\Common\util\lzw.cpp" />
     <ClCompile Include="..\..\Common\util\memorystream.cpp" />
     <ClCompile Include="..\..\Common\util\path.cpp" />
-    <ClCompile Include="..\..\Common\util\path_ex.cpp" />
     <ClCompile Include="..\..\Common\util\stdio_compat.c" />
     <ClCompile Include="..\..\Common\util\stream.cpp" />
     <ClCompile Include="..\..\Common\util\string.cpp" />
