+++ Solutions/Common.Lib/Common.Lib.vcxproj.filters
@@ -482,9 +482,6 @@
     <ClCompile Include="..\..\Common\util\memorystream.cpp">
       <Filter>Source Files\util</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\Common\util\path_ex.cpp">
-      <Filter>Source Files\util</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\Common\ac\spritefile.cpp">
       <Filter>Source Files\ac</Filter>
     </ClCompile>
